+++ app2_client/assets/kakao_map.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>Kakao Map Example</title>
+    <!-- Kakao 지도 API 호출: YOUR_KAKAO_APP_KEY 부분을 실제 발급받은 키로 변경 -->
+    <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey=YOUR_KAKAO_APP_KEY"></script>
+    <style>
+        html, body {
+          width: 100%;
+          height: 100%;
+          margin: 0;
+          padding: 0;
+        }
+        #map {
+          width: 100%;
+          height: 100%;
+        }
+    </style>
+</head>
+<body>
+<div id="map"></div>
+<script>
+    // Flutter에서 전달받은 초기 좌표 값(플레이스홀더)
+    // 만약 값이 없으면 기본값으로 서울 시청 좌표(37.5665, 126.9780)를 사용
+    var initialLat = parseFloat("__INITIAL_LAT__") || 37.5665;
+    var initialLng = parseFloat("__INITIAL_LNG__") || 126.9780;
+
+    var container = document.getElementById('map');
+    var options = {
+      center: new kakao.maps.LatLng(initialLat, initialLng),
+      level: 3
+    };
+
+    // 지도 생성
+    var map = new kakao.maps.Map(container, options);
+
+    // 드래그 가능한 마커 생성 및 지도에 표시
+    var marker = new kakao.maps.Marker({
+      position: map.getCenter(),
+      draggable: true
+    });
+    marker.setMap(map);
+
+    // 드래그가 끝났을 때 호출되어, 선택된 마커의 위치를 JSON 문자열로 반환하는 함수
+    function getSelectedDestination() {
+      var pos = marker.getPosition();
+      return JSON.stringify({
+        lat: pos.getLat(),
+        lng: pos.getLng()
+      });
+    }
+</script>
+</body>
+</html>
