+++ app2_client/lib/screens/destination_map_screen.dart
@@ -6,7 +6,11 @@ import 'search_address_screen.dart';
 import 'destination_radius_screen.dart';
 
 class DestinationMapScreen extends StatefulWidget {
-  const DestinationMapScreen({Key? key}) : super(key: key);
+  // 선택한 위치의 좌표를 전달받도록 수정 (optional)
+  final double? initialLat;
+  final double? initialLng;
+
+  const DestinationMapScreen({Key? key, this.initialLat, this.initialLng}) : super(key: key);
 
   @override
   _DestinationMapScreenState createState() => _DestinationMapScreenState();
@@ -24,7 +28,12 @@ class _DestinationMapScreenState extends State<DestinationMapScreen> {
 
   Future<void> _loadLocalHtml() async {
     _localHtml = await rootBundle.loadString('assets/kakao_map.html');
-    // _controller가 이미 생성된 경우 HTML 로드
+    // 만약 initialLat, initialLng 값이 있으면, HTML 파일 내 플레이스홀더를 대체
+    if (widget.initialLat != null && widget.initialLng != null) {
+      _localHtml = _localHtml
+          .replaceAll('__INITIAL_LAT__', widget.initialLat.toString())
+          .replaceAll('__INITIAL_LNG__', widget.initialLng.toString());
+    }
     if (_controller != null && _localHtml.isNotEmpty) {
       _controller.loadHtmlString(_localHtml);
     }
@@ -40,22 +49,21 @@ class _DestinationMapScreenState extends State<DestinationMapScreen> {
             javascriptMode: JavascriptMode.unrestricted,
             onWebViewCreated: (controller) {
               _controller = controller;
-              // 이미 _localHtml이 로드되었으면 바로 HTML을 로드
               if (_localHtml.isNotEmpty) {
                 _controller.loadHtmlString(_localHtml);
               }
             },
           ),
           // 중앙에 고정된 위치 마커 (선택된 위치 표시)
-          Center(
+          const Center(
             child: Icon(
               Icons.location_on,
               size: 50,
               color: Colors.redAccent,
             ),
           ),
-          // 왼쪽 상단: CustomBackButton (재사용 가능한 뒤로가기 버튼)
-          Positioned(
+          // 왼쪽 상단: CustomBackButton
+          const Positioned(
             top: 40,
             left: 16,
             child: CustomBackButton(),
@@ -73,10 +81,8 @@ class _DestinationMapScreenState extends State<DestinationMapScreen> {
                 ),
               ),
               onPressed: () async {
-                // 로컬 HTML 내에 정의된 JavaScript 함수 호출
                 String result = await _controller.evaluateJavascript("getSelectedDestination()");
                 print('선택된 위치 정보: $result');
-                // 이후, 결과에 따라 다음 페이지로 이동하거나 추가 처리를 함
                 Navigator.push(
                   context,
                   MaterialPageRoute(
